Weather icon outputted and show keyboard visible for city change
