+++ app/src/main/res/layout/calendar_main.xml
@@ -83,7 +83,7 @@
         android:layout_height="wrap_content"
         android:layout_centerVertical="true"
         android:layout_centerHorizontal="true"
-        android:textColor="#FFFFFF"
+        android:textColor="@color/colorPrimary"
         android:textAppearance="?android:attr/textAppearanceLarge"
         android:textSize="90sp" />
